- strategies. First steps.

+++ Candle.cs
@@ -9,18 +9,18 @@ namespace CandleTimeSeriesAnalysis
         public DateTime Start { get; set; }
         public TimeSpan Duration { get; set; }
         public DateTime End => Start.Add ( Duration );
-        public decimal Max { get; set; }
-        public decimal Min { get; set; }
-        public decimal Open { get; set; }
-        public decimal Close { get; set; }
-        public decimal BuyVolume { get; set; }
-        public decimal SellVolume { get; set; }
-        public decimal Volume => BuyVolume + SellVolume;
+        public double Max { get; set; }
+        public double Min { get; set; }
+        public double Open { get; set; }
+        public double Close { get; set; }
+        public double BuyVolume { get; set; }
+        public double SellVolume { get; set; }
+        public double Volume => BuyVolume + SellVolume;
         public bool GoesUp => Open <= Close;
         public bool GoesDown => !GoesUp;
-        public decimal Body => Math.Abs ( Open - Close );
-        public decimal Range => Max - Min;
-        public decimal BodyRangeRatio => Body / (Range > 0 ? Range : 1);
+        public double Body => Math.Abs ( Open - Close );
+        public double Range => Max - Min;
+        public double BodyRangeRatio => Body / (Range > 0 ? Range : 1);
 
         public static Candle FromTrades( IEnumerable<Trade> trades )
         {
