+++ Strategies/IWallet.cs
@@ -3,10 +3,10 @@
     public interface IWallet
     {
         string StockId { get; }
-        decimal Share { get; }
-        decimal Liquidity { get; }
-        void Sell( decimal share, decimal price );
-        void Buy( decimal share, decimal price );
-        void Pay( decimal price );
+        double Share { get; }
+        double Liquidity { get; }
+        void Sell( double share, double price );
+        void Buy( double share, double price );
+        void Pay( double price );
     }
 }
