+++ Strategies/Wallet.cs
@@ -3,27 +3,27 @@
     public class Wallet : IWallet
     {
         public string StockId { get; }
-        public decimal Share { get; private set; }
-        public decimal Liquidity { get; private set; }
+        public double Share { get; private set; }
+        public double Liquidity { get; private set; }
 
-        public Wallet( string stockId, decimal share, decimal liquidity )
+        public Wallet( string stockId, double share, double liquidity )
         {
             StockId = stockId;
             Share = share;
             Liquidity = liquidity;
         }
 
-        public void Sell( decimal share, decimal price )
+        public void Sell( double share, double price )
         {
             Share -= share;
             Liquidity += share * price;
         }
-        public void Buy( decimal share, decimal price )
+        public void Buy( double share, double price )
         {
             Share += share;
             Liquidity -= share * price;
         }
-        public void Pay( decimal price )
+        public void Pay( double price )
         {
             Liquidity -= price;
         }
